+++ ez.ai/client/src/components/ChatbotBuild/Preview/BasicPreview/LocationPreview.js
@@ -1,0 +1,34 @@
+import React from "react"
+import GoogleMapPresenter from "../../GoogleMapPresenter";
+
+const LocationPreview = ({
+  v,
+  i,
+  setClickedMainInput,
+  setNow,
+  onDelete
+}) => {
+  return(
+    <div className="main-content main-locabox" key={v.content + i}>
+      {" "}
+      <div
+        onClick={() => {
+          setClickedMainInput(v);
+          setNow(i);
+        }}
+      >
+        <GoogleMapPresenter />
+      </div>
+      <div
+        className="tool-delete delete-location"
+        onClick={() => {
+          onDelete(v.id);
+        }}
+      >
+        <i className="fas fa-times"></i>
+      </div>
+    </div>
+  );
+};
+
+export default LocationPreview;
